Update BOSH and YQ
This updates bosh and yq versions and also optimises how image layers
are arranged.

For instance when it does apt-get install xxx in one RUN and `apt-get
clean && rm -rf /var/lib/apt/lists/*` in another one it actually does
not decrease the size of the image as it creates two layers - one with
downloaded packages and another one with deleted data.